se implemento tabla para admin con nextui y se adapto a lo que ya habia antes
