+++ client/src/views/Admin/Admin.jsx
@@ -24,8 +24,8 @@ function classNames(...classes) {
 export default function Admin() {
   return (
     <>
-      <div className="min-h-full min-w-full flex-1">
-        <Disclosure as="nav" className="bg-[#1F0A4D]">
+      <div className="min-h-full min-w-full flex-1 flex flex-col items-center">
+        <Disclosure as="nav" className="bg-[#1F0A4D] w-full">
           {({ open }) => (
             <>
               <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
@@ -188,11 +188,11 @@ export default function Admin() {
             </>
           )}
         </Disclosure>
-      </div>
-      <Routes>
-        <Route path="/" element={<AdminHome />} />
-        <Route path="/create" element={<Create />} />
+        <Routes>
+          <Route path="/" element={<AdminHome/>} />
+          <Route path="/create" element={<Create />} />
       </Routes>
+      </div>
     </>
   );
 }
