+++ client/src/views/Admin/AdminHome/AdminHome.jsx
@@ -1,10 +1,9 @@
-import ListGames from "../ListGames/ListGames";
 import Filters from "../../../components/Filters/Filters";
 import Pagination from "../../../components/Pagination/Pagination";
 import { useSelector, useDispatch } from "react-redux";
 import { getGames } from "../../../redux/actions";
 import { useState, useEffect } from "react";
-import GamesTable from "../Elementgame/GamesTable";
+import GamesTable from "../GamesTable/GamesTable";
 
 const AdminHome = () => {
   const dispatch = useDispatch();
@@ -28,7 +27,6 @@ const AdminHome = () => {
   };
 
   useEffect(() => {
-    console.log("Actualio estado!!!");
     dispatch(getGames(filters));
   }, [filters, dispatch]);
 
@@ -40,8 +38,7 @@ const AdminHome = () => {
       <div className="flex-1 h-auto">
         {games.videogames ? (
           <>
-            {/* <ListGames videogames={games.videogames} /> */}
-            <GamesTable/>
+            <GamesTable videogames={games.videogames}/>
             <Pagination
               PaginationData={games.PaginationData}
               onPageChange={onPageChangeHandler}
