--- client/src/views/Admin/ListGames/ListGames.jsx
@@ -1,17 +1,0 @@
-import ElementGame from "../Elementgame/Elementgame";
-
-const ListGames = ({ videogames }) => {
-
-  return (
-    <div>
-      <ul>
-        {videogames.map((game) => (
-          <li key={game.id}>
-            <ElementGame game={game} />
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-};
-export default ListGames;
